Create card view
Allows for showing each card one at a time,
and selecting an answer, and seeing the answer.

Each card has some metadata - labels, a difficulty, and
completion status.

The card and metadata are editable. Cards can be deleted,
though as of yet not created.

Currently, some 792 initial cards downloaded from a revision
website for the Irish driving theory test are hardcoded in
for testing. There is no saving/loading.